23 type conversions using "as" and "std::convert" module

+++ exercises/23_conversions/1_using_as.rs
@@ -10,11 +10,9 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64 // Total is f64, so we need to cast values.len() (type usize) to f64
 }
 
 fn main() {
